controllers work

+++ server/controllers/programControllers.js
@@ -14,7 +14,7 @@ exports.createProgram = async (req, res) => {
     rating,
   } = req.body;
 
-  if ((!name || !bio, !website_url, !borough, !img_url, !color)) return res.sendStatus(400);
+  if (!name || !bio || !website_url || !borough || !img_url || !color) return res.sendStatus(400);
 
   const isAvailable = (await Program.findByName(name)) === null;
   if (!isAvailable) return res.sendStatus(400);
@@ -49,7 +49,9 @@ exports.updateProgram = async (req, res) => {
   const { id } = req.params;
   const { organization_id, name, bio, website_url, borough, img_url, color } = req.body;
 
-  if (!isAuthorized(organization_id, req.session)) return res.sendStatus(403);
+  const program = await Program.findById(id);
+  if (program === null) return res.sendStatus(404);
+  if (Number(program.organizationId) !== Number(req.session.organizationId)) return res.sendStatus(403);
 
   const updatedProgram = await Program.update(id, name, bio, website_url, borough, img_url, color);
   if (updatedProgram === null) return res.sendStatus(404);
@@ -64,7 +66,6 @@ exports.listAllPrograms = async (req,res) => {
 exports.getRecommends = async (req, res) => {
   const { id } = req.params;
   const recommends = await Program.getRecommends(id);
-  console.log(recommends);
   res.send(recommends);
 };
 
@@ -79,6 +80,7 @@ exports.deleteProgram = async (req,res) => {
   const {id} = req.params;
   const isAvailable = await Program.findById(id);
   if (isAvailable === null) return res.sendStatus(404);
+  if (Number(isAvailable.organizationId) !== Number(req.session.organizationId)) return res.sendStatus(403);
 
   const comments = await knex.raw('DELETE FROM comments WHERE program_id = ?', [id]);
   const recommends = await knex.raw('DELETE FROM recommends WHERE program_id = ?', [id]);
